backend/requirements.txt
fastapi~=0.75.1
uvicorn[standard]~=0.17.6

mysql-connector-python

pydantic~=1.9.0
python-dotenv==0.20.0